Still having a problem connecting wordpress and mariadb

+++ srcs/requirements/wordpress/tools/wordpress.sh
@@ -1,27 +1,31 @@
 #!/bin/bash
 
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+sed -i '36 s/\/run\/php\/php7.3-fpm.sock/9000/' /etc/php/7.3/fpm/pool.d/www.conf
 
-chmod +x wp-cli.phar
+if [ ! -f /var/www/html/wp-config.php ]; then
+	curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
 
-mv wp-cli.phar /usr/bin/wp
+	chmod +x wp-cli.phar
 
-chown -R www-data:www-data /var/www/html/
-chmod -R 755 /var/www/html
+	mv wp-cli.phar /usr/bin/wp
 
-cd /var/www/html
+	chown -R www-data:www-data /var/www/html/
+	chmod -R 755 /var/www/html
 
-wp core download --allow-root
+	cd /var/www/html
 
-cp wp-config-sample.php wp-config.php
+	wp core download --allow-root
 
-wp config set DB_NAME $MYSQL_DATABASE --allow-root
-wp config set DB_USER $MYSQL_USER --allow-root
-wp config set DB_PASSWORD $MYSQL_PASSWORD --allow-root
-wp config set DB_HOST $MYSQL_HOST --allow-root
+	cp wp-config-sample.php wp-config.php
 
-wp core install --url=$DOMAIN_NAME --title="WordPress Site" --admin_user=$ADMIN_USER --admin_password=$ADMIN_PASSWORD --admin_email=$ADMIN_EMAIL --allow-root
+	wp config set DB_NAME $MYSQL_DATABASE --allow-root
+	wp config set DB_USER $MYSQL_USER --allow-root
+	wp config set DB_PASSWORD $MYSQL_PASSWORD --allow-root
+	wp config set DB_HOST $MYSQL_HOST --allow-root
 
-wp user create $USER $USER_EMAIL --user_pass=$USER_PASSWORD role='author' --allow-root
+	wp core install --url=$DOMAIN_NAME --title="WordPress Site" --admin_user=$ADMIN_USER --admin_password=$ADMIN_PASSWORD --admin_email=$ADMIN_EMAIL --allow-root
+
+	wp user create $USER $USER_EMAIL --user_pass=$USER_PASSWORD role='author' --allow-root
+fi
 
 exec "$@"
